Reorganize ui-kit directory
