+++ app/components/map-panel/map-panel-selfie-uploaded-by-target.js
@@ -5,7 +5,7 @@ import {
 } from 'react-native';
 
 import MapPanelStyles from './map-panel-styles';
-import { H2, Caption2 } from '../../components/ui-kit/typography';
+import { H2, Caption2 } from '../../components/ui-kit/atoms/typography';
 import DaterButton from '../../components/ui-kit/atoms/dater-button';
 import cloudinaryUrl from '../../utils/cloudinary-utils';
