+++ app/screens/profile/profile-screen.js
@@ -9,9 +9,9 @@ import {
 } from 'react-native';
 import ImageLoad from 'react-native-image-placeholder';
 
-import DaterModal from '../../components/ui-kit/dater-modal';
+import DaterModal from '../../components/ui-kit/organisms/dater-modal';
 import DaterButton from '../../components/ui-kit/atoms/dater-button';
-import { H2 } from '../../components/ui-kit/typography';
+import { H2 } from '../../components/ui-kit/atoms/typography';
 import cloudinaryUrl from '../../utils/cloudinary-utils';
 import { PrivateUserRecord } from '../../types';
 import { SCREEN_WIDTH } from '../../constants';
