+++ app/screens/profile/upload-photo-screen.js
@@ -4,9 +4,9 @@ import {
   Image,
 } from 'react-native';
 
-import DaterModal from '../../components/ui-kit/dater-modal';
+import DaterModal from '../../components/ui-kit/organisms/dater-modal';
 import DaterButton from '../../components/ui-kit/atoms/dater-button';
-import { H2, Body } from '../../components/ui-kit/typography';
+import { H2, Body } from '../../components/ui-kit/atoms/typography';
 
 const uploadPhotoIcon = require('../../assets/icons/upload-photo/upload-photo.png');
