+++ app/screens/ui-kit/textinputs-screen.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { ScrollView } from 'react-native';
 
 import DaterTextInput from '../../components/ui-kit/atoms/dater-text-input';
-import DaterModal from '../../components/ui-kit/dater-modal';
+import DaterModal from '../../components/ui-kit/organisms/dater-modal';
 import DaterButton from '../../components/ui-kit/atoms/dater-button';
 
 type Props = {
